client/requirements.txt
requests
mss
pynput
pywin32
cryptography
Pillow
psutil
